Revert "feat: Implemented configurable task management and improved AI processing."
This reverts commit cbb4b6fe619287b17f2e9d5b70924e8d06681602.
