+++ Processor.test.ts
@@ -1,15 +1,11 @@
 import { Mocks } from './Mocks';
 import { Processor } from './Processor';
 import { AIAnalyzer } from './AIAnalyzer';
-import { TasksManagerFactory } from './TasksManagerFactory';
-import { GoogleTasksManager } from './GoogleTasksManager';
-import { TodoistManager } from './TodoistManager';
+import { TasksManager } from './TasksManager';
 import { Config } from './Config';
 
 jest.mock('./AIAnalyzer');
-jest.mock('./TasksManagerFactory');
-jest.mock('./GoogleTasksManager');
-jest.mock('./TodoistManager');
+jest.mock('./TasksManager');
 jest.mock('./Config');
 
 global.SpreadsheetApp = {
@@ -27,20 +23,15 @@ global.Logger = {
 };
 
 describe('Processor Tests', () => {
-  let mockGoogleTasksManager: jest.Mocked<GoogleTasksManager>;
-  let mockTodoistManager: jest.Mocked<TodoistManager>;
-
   beforeEach(() => {
     // Reset mocks before each test
     (AIAnalyzer.generatePlan as jest.Mock).mockClear();
-    (TasksManagerFactory.getTasksManager as jest.Mock).mockClear();
+    (TasksManager.findCheckpoint as jest.Mock).mockClear();
+    (TasksManager.upsertTask as jest.Mock).mockClear();
     (Config.getConfig as jest.Mock).mockClear();
-
-    mockGoogleTasksManager = new (GoogleTasksManager as any)();
-    mockTodoistManager = new (TodoistManager as any)();
   });
 
-  it('should process a simple thread correctly with Google Tasks', () => {
+  it('should process a simple thread correctly', () => {
     // Setup
     const mockThread = Mocks.getMockThread({
       getId: () => 'thread-1',
@@ -52,20 +43,121 @@ describe('Processor Tests', () => {
     });
 
     const mockConfig = {
-      task_service: 'Google Tasks',
       unprocessed_label: 'unprocessed',
       processed_label: 'processed',
+      processing_failed_label: 'error',
       max_threads: 50,
     };
     (Config.getConfig as jest.Mock).mockReturnValue(mockConfig);
-    (TasksManagerFactory.getTasksManager as jest.Mock).mockReturnValue(mockGoogleTasksManager);
 
     const mockPlan = {
       action: { move_to: 'ARCHIVE', mark_read: true },
       task: { title: 'Test Task', notes: 'Test Notes' },
     };
     (AIAnalyzer.generatePlan as jest.Mock).mockReturnValue(mockPlan);
-    (mockGoogleTasksManager.findCheckpoint as jest.Mock).mockReturnValue(null);
+    (TasksManager.findCheckpoint as jest.Mock).mockReturnValue(null);
+
+    const unprocessedLabel = { getThreads: () => [mockThread] };
+    const processedLabel = { getName: () => 'processed' };
+    global.GmailApp = {
+      getUserLabelByName: jest.fn((name: string) => {
+        if (name === 'unprocessed') {
+          return unprocessedLabel as any;
+        }
+        if (name === 'processed') {
+            return processedLabel as any;
+        }
+        return {
+          getName: () => name,
+        } as any;
+      }),
+    } as any;
+
+    // Execute
+    Processor.processAllUnprocessedThreads();
+
+    // Verify
+    expect(TasksManager.upsertTask).toHaveBeenCalled();
+    expect(mockThread.addLabel).toHaveBeenCalled();
+  });
+
+  it('should use the email subject as a fallback title', () => {
+    // Setup
+    const mockThread = Mocks.getMockThread({
+      getId: () => 'thread-1',
+      getFirstMessageSubject: () => 'Fallback Subject',
+      getMessages: () => [Mocks.getMockMessage({ getDate: () => new Date() })],
+      removeLabel: jest.fn(),
+      addLabel: jest.fn(),
+      moveToArchive: jest.fn(),
+    });
+
+    const mockConfig = {
+      unprocessed_label: 'unprocessed',
+      processed_label: 'processed',
+      processing_failed_label: 'error',
+      max_threads: 50,
+    };
+    (Config.getConfig as jest.Mock).mockReturnValue(mockConfig);
+
+    const mockPlan = {
+      action: { move_to: 'ARCHIVE', mark_read: true },
+      task: { title: '', notes: 'Test Notes' },
+    };
+    (AIAnalyzer.generatePlan as jest.Mock).mockReturnValue(mockPlan);
+    (TasksManager.findCheckpoint as jest.Mock).mockReturnValue(null);
+
+    const unprocessedLabel = { getThreads: () => [mockThread] };
+    const processedLabel = { getName: () => 'processed' };
+    global.GmailApp = {
+        getUserLabelByName: jest.fn((name: string) => {
+            if (name === 'unprocessed') {
+                return unprocessedLabel as any;
+            }
+            if (name === 'processed') {
+                return processedLabel as any;
+            }
+            return {
+                getName: () => name,
+            } as any;
+        }),
+    } as any;
+
+    // Execute
+    Processor.processAllUnprocessedThreads();
+
+    // Verify
+    expect(TasksManager.upsertTask).toHaveBeenCalledWith(
+        mockThread,
+        { title: 'Fallback Subject', notes: 'Test Notes' },
+        mockConfig
+    );
+  });
+
+  it('should not create a task and mark as unread if the email is not actionable', () => {
+    // Setup
+    const mockThread = Mocks.getMockThread({
+      getId: () => 'thread-2',
+      getFirstMessageSubject: () => 'Non-Actionable Email',
+      getMessages: () => [Mocks.getMockMessage({ getDate: () => new Date() })],
+      removeLabel: jest.fn(),
+      addLabel: jest.fn(),
+      markUnread: jest.fn(),
+      moveToArchive: jest.fn(),
+    });
+
+    const mockConfig = {
+      unprocessed_label: 'unprocessed',
+      processed_label: 'processed',
+      max_threads: 50,
+    };
+    (Config.getConfig as jest.Mock).mockReturnValue(mockConfig);
+
+    const mockPlan = {
+      action: { move_to: 'ARCHIVE', mark_read: true }, // This will be overridden
+    };
+    (AIAnalyzer.generatePlan as jest.Mock).mockReturnValue(mockPlan);
+    (TasksManager.findCheckpoint as jest.Mock).mockReturnValue(null);
 
     const unprocessedLabel = { getThreads: () => [mockThread] };
     const processedLabel = { getName: () => 'processed' };
@@ -81,52 +173,8 @@ describe('Processor Tests', () => {
     Processor.processAllUnprocessedThreads();
 
     // Verify
-    expect(mockGoogleTasksManager.upsertTask).toHaveBeenCalled();
-    expect(mockThread.addLabel).toHaveBeenCalled();
-  });
-
-  it('should process a simple thread correctly with Todoist', () => {
-    // Setup
-    const mockThread = Mocks.getMockThread({
-      getId: () => 'thread-1',
-      getFirstMessageSubject: () => 'Test Subject',
-      getMessages: () => [Mocks.getMockMessage({ getDate: () => new Date() })],
-      removeLabel: jest.fn(),
-      addLabel: jest.fn(),
-      moveToArchive: jest.fn(),
-    });
-
-    const mockConfig = {
-      task_service: 'Todoist',
-      unprocessed_label: 'unprocessed',
-      processed_label: 'processed',
-      max_threads: 50,
-    };
-    (Config.getConfig as jest.Mock).mockReturnValue(mockConfig);
-    (TasksManagerFactory.getTasksManager as jest.Mock).mockReturnValue(mockTodoistManager);
-
-    const mockPlan = {
-      action: { move_to: 'ARCHIVE', mark_read: true },
-      task: { title: 'Test Task', notes: 'Test Notes' },
-    };
-    (AIAnalyzer.generatePlan as jest.Mock).mockReturnValue(mockPlan);
-    (mockTodoistManager.findCheckpoint as jest.Mock).mockReturnValue(null);
-
-    const unprocessedLabel = { getThreads: () => [mockThread] };
-    const processedLabel = { getName: () => 'processed' };
-    global.GmailApp = {
-      getUserLabelByName: jest.fn((name: string) => {
-        if (name === 'unprocessed') return unprocessedLabel as any;
-        if (name === 'processed') return processedLabel as any;
-        return { getName: () => name } as any;
-      }),
-    } as any;
-
-    // Execute
-    Processor.processAllUnprocessedThreads();
-
-    // Verify
-    expect(mockTodoistManager.upsertTask).toHaveBeenCalled();
-    expect(mockThread.addLabel).toHaveBeenCalled();
+    expect(TasksManager.upsertTask).not.toHaveBeenCalled();
+    expect(mockThread.markUnread).toHaveBeenCalled();
+    expect(mockThread.moveToArchive).toHaveBeenCalled();
   });
 });
